implemented all history capabilities and fixed searchbar onFocus

+++ front-dispatcher/src/helpers/localStorageUsage.ts
@@ -1,6 +1,6 @@
 import { historyStrings } from "../strings/strings";
 
-export const getHistory = () => {
+export const getHistory = (): [string] => {
   const storageData = localStorage.getItem(historyStrings.History);
   return storageData ? JSON.parse(storageData) : [];
 };
@@ -9,11 +9,18 @@ export const ClearSearchHistory = () => {
 };
 export const addToHistory = (recentSearch: string) => {
   const storageData = localStorage.getItem(historyStrings.History);
-  const savedStorage = storageData ? JSON.parse(storageData) : [];
+  const savedStorage: [string] = storageData ? JSON.parse(storageData) : [];
+  if (isAlreadyInStorage(recentSearch)) return;
   localStorage.setItem(
     historyStrings.History,
     JSON.stringify([...savedStorage, recentSearch])
   );
+};
+export const isAlreadyInStorage = (recentSearch: string) => {
+  const storageData = localStorage.getItem(historyStrings.History);
+  if (!storageData) return false;
+  const savedStorage: [string] = JSON.parse(storageData);
+  return savedStorage.includes(recentSearch);
 };
 export const clearSearchFromHistory = (searchToRemove: string) => {
   const storageData = localStorage.getItem(historyStrings.History);
